+++ app/ts/library/typescript.ts
@@ -5,7 +5,8 @@ export type ObjectTupleToValueTuple<T extends readonly [...readonly {}[]], K ext
 export type ObjectUnionToKeyedObjectUnion<T extends {}, K extends keyof T> = T extends {} ? T[K] extends PropertyKey ? { [_ in T[K]]: T } : never : never
 export type UnionToIntersection<T> = (T extends unknown ? (k: T) => void : never) extends (k: infer I) => void ? I : never
 
-export type NarrowUnion<T, Discriminant> = T extends Discriminant ? T : never
+export type PermitUnion<T, Discriminant> = T extends Discriminant ? T : never
+export type OmitUnion<T, Discriminant> = T extends Discriminant ? never : T
 
 export type ResolvePromise<T> = T extends PromiseLike<infer R> ? R : never
